Move constants and utils to higher level directories
