+++ client/dashboard/me/purchases/purchase-payment-method.tsx
@@ -1,6 +1,11 @@
 import { __ } from '@wordpress/i18n';
+import {
+	isAutoRenewEnabled,
+	isExpired,
+	isRenewing,
+	isAkismetFreeProduct,
+} from '../../utils/purchase';
 import { PaymentMethodImage } from './payment-method-image';
-import { isAutoRenewEnabled, isExpired, isRenewing, isAkismetFreeProduct } from './util';
 import type { Purchase } from '../../data/purchase';
 
 export function PurchasePaymentMethod( {
